# Use NTT form for DB multiplications