@yeungbo 2021-10-19
I noticed that  the Time and CPU are all increased (it spends more time, right?) in the After benchmark than Before, why it is said the improvement has ~3x speedup.....