@kshehata 2021-11-01
You're misreading the table. I'll make this simple.

Before:
`PIRFixture/ServerProcessRequest/65536  2216483948 ns   1771607000 ns            1`

After:
`PIRFixture/ServerProcessRequest/65536   552863281 ns    537544000 ns            1`

What this means is that the ProcessRequest function on the Server, when given a payload size of 65536 bytes, went from taking 2.2 seconds to 0.55 seconds. That's a speedup of ~4x.