@yeungbo 2021-11-02
ah, I see. It is only for speeding up the performance of ProcessRequest on the Server. Thank you for your clarifications!